Added: Question 4

+++ src/patterns/LeftSkewedBinaryTriangle.java
@@ -13,12 +13,12 @@ package patterns;
 
 public class LeftSkewedBinaryTriangle {
     public static void main(String[] args) {
-        int num = 5;
+        byte num = 5;
 
         outerLoop:
-        for (int i = 0; i < num; i++) {
+        for (byte i = 0; i < num; i++) {
             innerLoop:
-            for (int j = 0; j < num-i; j++) {
+            for (byte j = 0; j < num-i; j++) {
                 if (i % 2 == 1){
                     System.out.print("0");
                 }
