+++ src/patterns/LeftSkewedStarTriangle.java
@@ -13,10 +13,10 @@ package patterns;
 
 public class LeftSkewedStarTriangle {
     public static void main(String[] args) {
-        int num = 5;
+        byte num = 5;
 
-        for (int i = 0; i < num; i++) {
-            for (int j = 0; j < num-i; j++) {
+        for (byte i = 0; i < num; i++) {
+            for (byte j = 0; j < num-i; j++) {
                 System.out.print("*");
             }
             System.out.println();
